refactor(agent-infra): mcp-client esm to cjs
