+++ apps/agent-tars/src/renderer/package.json
@@ -45,7 +45,7 @@
     "path-browserify": "1.0.1"
   },
   "devDependencies": {
-    "@modelcontextprotocol/sdk": "^1.10.0",
+    "@modelcontextprotocol/sdk": "^1.10.2",
     "typescript": "^5.7.3",
     "@types/react": "^18.3.18",
     "@types/react-dom": "^18.3.5",
