+++ packages/agent-infra/mcp-client/package.json
@@ -3,15 +3,14 @@
   "version": "0.4.3",
   "description": "An MCP Client to run servers for Electron apps, support same-process approaching",
   "private": false,
-  "type": "module",
   "types": "./dist/index.d.ts",
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.js",
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
   "files": [
     "dist"
   ],
   "scripts": {
-    "dev": "tsx src/test.ts",
+    "dev": "tsx examples/test.ts",
     "clean": "rm -rf dist",
     "build": "rslib build",
     "prepare": "npm run build",
@@ -21,7 +20,7 @@
     "test:integration": "vitest tests/integration"
   },
   "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.10.0",
+    "@modelcontextprotocol/sdk": "^1.10.2",
     "@agent-infra/mcp-shared": "workspace:*",
     "zod": "^3.23.8",
     "uuid": "^11.1.0"
