+++ packages/agent-infra/mcp-servers/browser/package.json
@@ -22,7 +22,7 @@
     "watch": "rslib build --watch"
   },
   "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.10.0"
+    "@modelcontextprotocol/sdk": "^1.10.2"
   },
   "devDependencies": {
     "@agent-infra/logger": "workspace:*",
