+++ packages/agent-infra/mcp-servers/commands/package.json
@@ -24,7 +24,7 @@
     "test:integration": "vitest tests/integration"
   },
   "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.10.0"
+    "@modelcontextprotocol/sdk": "^1.10.2"
   },
   "devDependencies": {
     "tsx": "^4.19.3",
