+++ packages/agent-infra/mcp-servers/filesystem/package.json
@@ -20,7 +20,7 @@
     "watch": "rslib build --watch"
   },
   "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.10.0"
+    "@modelcontextprotocol/sdk": "^1.10.2"
   },
   "devDependencies": {
     "diff": "^5.1.0",
