+++ packages/agent-infra/mcp-shared/package.json
@@ -35,7 +35,7 @@
     "watch": "rslib build --watch"
   },
   "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.10.0",
+    "@modelcontextprotocol/sdk": "^1.10.2",
     "zod": "^3.23.8"
   },
   "devDependencies": {
